Split machine.go to multiples files. One per public method.
