+++ pkg/crc/machine/stop.go
@@ -1,0 +1,38 @@
+package machine
+
+import (
+	"github.com/code-ready/crc/pkg/crc/logging"
+	"github.com/code-ready/machine/libmachine/state"
+	"github.com/pkg/errors"
+)
+
+func (client *client) Stop() (state.State, error) {
+	defer unsetMachineLogging()
+
+	// Set libmachine logging
+	err := setMachineLogging(client.debug)
+	if err != nil {
+		return state.None, errors.Wrap(err, "Cannot initialize logging")
+	}
+
+	libMachineAPIClient, cleanup, err := createLibMachineClient(client.debug)
+	defer cleanup()
+	if err != nil {
+		return state.None, errors.Wrap(err, "Cannot initialize libmachine")
+	}
+	host, err := libMachineAPIClient.Load(client.name)
+
+	if err != nil {
+		return state.None, errors.Wrap(err, "Cannot load machine")
+	}
+
+	logging.Info("Stopping the OpenShift cluster, this may take a few minutes...")
+	if err := host.Stop(); err != nil {
+		status, err := host.Driver.GetState()
+		if err != nil {
+			logging.Debugf("Cannot get VM status after stopping it: %v", err)
+		}
+		return status, errors.Wrap(err, "Cannot stop machine")
+	}
+	return host.Driver.GetState()
+}
